Extracted XtextStyledTextSelectionProvider from StyledTextXtextAdapter to its own class. InlineDefinitionSectionStyle is now set to false by default for new diagrams. Introduce StatechartDefinitionSection - extracted and reworked from StatechartDiagramEditor. fixed wrong package in plugin.xml
